+++ exporters/otlp/test/otlp_grpc_metric_exporter_test.cc
@@ -27,6 +27,9 @@
 
 #  include "opentelemetry/exporters/otlp/protobuf_include_suffix.h"
 
+#  include "opentelemetry/exporters/otlp/otlp_grpc_client.h"
+#  include "opentelemetry/exporters/otlp/otlp_grpc_client_factory.h"
+
 #  include "opentelemetry/sdk/trace/simple_processor.h"
 #  include "opentelemetry/sdk/trace/tracer_provider.h"
 #  include "opentelemetry/trace/provider.h"
@@ -52,10 +55,32 @@ class OtlpGrpcMetricExporterTestPeer : public ::testing::Test
 {
 public:
   std::unique_ptr<sdk::metrics::PushMetricExporter> GetExporter(
-      std::unique_ptr<proto::collector::metrics::v1::MetricsService::StubInterface> &stub_interface)
+      const OtlpGrpcMetricExporterOptions &options)
+  {
+    return std::unique_ptr<sdk::metrics::PushMetricExporter>(new OtlpGrpcMetricExporter(options));
+  }
+
+  std::unique_ptr<sdk::metrics::PushMetricExporter> GetExporter(
+      std::unique_ptr<proto::collector::metrics::v1::MetricsService::StubInterface> stub_interface)
   {
     return std::unique_ptr<sdk::metrics::PushMetricExporter>(
         new OtlpGrpcMetricExporter(std::move(stub_interface)));
+  }
+
+  std::unique_ptr<sdk::metrics::PushMetricExporter> GetExporter(
+      std::unique_ptr<proto::collector::metrics::v1::MetricsService::StubInterface> stub_interface,
+      std::shared_ptr<OtlpGrpcClient> client)
+  {
+    return std::unique_ptr<sdk::metrics::PushMetricExporter>(
+        new OtlpGrpcMetricExporter(std::move(stub_interface), std::move(client)));
+  }
+
+  std::unique_ptr<sdk::metrics::PushMetricExporter> GetExporter(
+      const OtlpGrpcMetricExporterOptions &options,
+      std::shared_ptr<OtlpGrpcClient> client)
+  {
+    return std::unique_ptr<sdk::metrics::PushMetricExporter>(
+        new OtlpGrpcMetricExporter(options, std::move(client)));
   }
 
   // Get the options associated with the given exporter.
@@ -205,6 +230,35 @@ TEST_F(OtlpGrpcMetricExporterTestPeer, ConfigUnknownInsecureFromEnv)
   unsetenv("OTEL_EXPORTER_OTLP_METRICS_INSECURE");
 }
 #  endif
+
+TEST_F(OtlpGrpcMetricExporterTestPeer, CheckGetAggregationTemporality)
+{
+  auto options                    = OtlpGrpcMetricExporterOptions();
+  options.aggregation_temporality = PreferredAggregationTemporality::kCumulative;
+
+  auto client = OtlpGrpcClientFactory::Create(options);
+
+  auto exporter0 = GetExporter(options);
+  auto exporter1 = GetExporter(client->MakeMetricsServiceStub());
+  auto exporter2 = GetExporter(options, client);
+  auto exporter3 = GetExporter(client->MakeMetricsServiceStub(), client);
+
+  EXPECT_EQ(
+      opentelemetry::sdk::metrics::AggregationTemporality::kCumulative,
+      exporter0->GetAggregationTemporality(opentelemetry::sdk::metrics::InstrumentType::kCounter));
+
+  EXPECT_EQ(
+      opentelemetry::sdk::metrics::AggregationTemporality::kCumulative,
+      exporter1->GetAggregationTemporality(opentelemetry::sdk::metrics::InstrumentType::kCounter));
+
+  EXPECT_EQ(
+      opentelemetry::sdk::metrics::AggregationTemporality::kCumulative,
+      exporter2->GetAggregationTemporality(opentelemetry::sdk::metrics::InstrumentType::kCounter));
+
+  EXPECT_EQ(
+      opentelemetry::sdk::metrics::AggregationTemporality::kCumulative,
+      exporter3->GetAggregationTemporality(opentelemetry::sdk::metrics::InstrumentType::kCounter));
+}
 
 }  // namespace otlp
 }  // namespace exporter